change VP_INT (int -> long)

+++ hos-v4/include/itron.h
@@ -54,7 +54,7 @@ typedef struct t_systim		/* システム時刻 */
 	UH utime;				/* 上位 16bit */
 } SYSTIM;
 
-typedef INT		VP_INT;		/* VP または INT */
+typedef long	VP_INT;		/* VP または INT */
 
 typedef ER		ER_BOOL;	/* ER または BOOL */
